Fix _sre buffer lifetime and migrate tests to apptest
interp_sre.py: make_ctx() acquires a buffer and releases it immediately
after the match context is built (single-shot search semantics match
CPython).  W_SRE_Scanner re-acquires the buffer and holds it for the
full iterator lifetime to prevent concurrent resize; it releases eagerly
on exhaustion, with _finalize_ as a safety net.

Tests: the large unit-test file test_app_sre.py is pared back to the
interp-level tests only; the app-level regex tests move to the new
apptest_sre.py (run under the translated interpreter).
