+++ pypy/module/_sre/test/test_app_sre.py
@@ -21,169 +21,6 @@ def init_app_test(cls, space):
         """)
 
 
-class AppTestSrePy:
-    def test_magic(self):
-        import _sre, sre_constants
-        assert sre_constants.MAGIC == _sre.MAGIC
-
-    def test_codesize(self):
-        import _sre
-        assert _sre.getcodesize() == _sre.CODESIZE
-        # in py3.7, it should always be 4
-        assert _sre.CODESIZE == 4
-
-    def test_opcodes(self):
-        import _sre
-        assert _sre.OPCODES[:4] == 'failure success any any_all'.split()
-
-
-class AppTestSrePattern:
-    def setup_class(cls):
-        # This imports support_test_sre as the global "s"
-        init_app_test(cls, cls.space)
-
-    spaceconfig = {'usemodules': ['itertools']}
-
-    def test_copy(self):
-        # new in 3.7
-        import re
-        p = re.compile("b")
-        assert p.__copy__() is p
-        assert p.__deepcopy__("whatever") is p
-
-    def test_creation_attributes(self):
-        import re
-        pattern_string = b"(b)l(?P<g>a)"
-        p = re.compile(pattern_string, re.I | re.M)
-        assert pattern_string == p.pattern
-        assert re.I | re.M == p.flags
-        assert 2 == p.groups
-        assert {"g": 2} == p.groupindex
-        raises(TypeError, "p.groupindex['g'] = 3")
-
-    def test_repeat_minmax_overflow(self):
-        import re
-        string = "x" * 100000
-        assert re.match(r".{%d}" % (self.s.MAXREPEAT - 1), string) is None
-        assert re.match(r".{,%d}" % (self.s.MAXREPEAT - 1), string).span() == (0, 100000)
-        assert re.match(r".{%d,}?" % (self.s.MAXREPEAT - 1), string) is None
-        import sys
-        if sys.version_info[:3] <= (3, 2, 3):
-            # XXX: These are fixed in 3.2.4 or so
-            return
-        raises(OverflowError, re.compile, r".{%d}" % self.s.MAXREPEAT)
-        raises(OverflowError, re.compile, r".{,%d}" % self.s.MAXREPEAT)
-        raises(OverflowError, re.compile, r".{%d,}?" % self.s.MAXREPEAT)
-
-    def test_match_none(self):
-        import re
-        p = re.compile("bla")
-        none_matches = ["b", "bl", "blub", "jupidu"]
-        for string in none_matches:
-            assert None == p.match(string)
-
-    def test_pos_endpos(self):
-        import re
-        # XXX maybe fancier tests here
-        p = re.compile("bl(a)")
-        tests = [("abla", 0, 4), ("abla", 1, 4), ("ablaa", 1, 4)]
-        for string, pos, endpos in tests:
-            assert p.search(string, pos, endpos)
-        tests = [("abla", 0, 3), ("abla", 2, 4)]
-        for string, pos, endpos in tests:
-            assert not p.search(string, pos, endpos)
-
-    def test_findall(self):
-        import re
-        assert ["b"] == re.findall("b", "bla")
-        assert ["a", "u"] == re.findall("b(.)", "abalbus")
-        assert [("a", "l"), ("u", "s")] == re.findall("b(.)(.)", "abalbus")
-        assert [("a", ""), ("s", "s")] == re.findall("b(a|(s))", "babs")
-        assert ['', '', 'X', '', ''] == re.findall("X??", "1X4")  # new in 3.7
-
-    def test_findall_unicode(self):
-        import re
-        assert [u"\u1234"] == re.findall(u"\u1234", u"\u1000\u1234\u2000")
-        assert ["a", "u"] == re.findall("b(.)", "abalbus")
-        assert [("a", "l"), ("u", "s")] == re.findall("b(.)(.)", "abalbus")
-        assert [("a", ""), ("s", "s")] == re.findall("b(a|(s))", "babs")
-        assert [(b"a", b""), (b"s", b"s")] == re.findall(b"b(a|(s))", b"babs")
-        assert [u"xyz"] == re.findall(u".*yz", u"xyz")
-
-    def test_finditer(self):
-        import re
-        it = re.finditer("b(.)", "brabbel")
-        assert "br" == next(it).group(0)
-        assert "bb" == next(it).group(0)
-        raises(StopIteration, next, it)
-
-    def test_split(self):
-        import re
-        assert ["a", "o", "u", ""] == re.split("b", "abobub")
-        assert ["a", "o", "ub"] == re.split("b", "abobub", 2)
-        assert ['', 'a', 'l', 'a', 'lla'] == re.split("b(a)", "balballa")
-        assert ['', 'a', None, 'l', 'u', None, 'lla'] == (
-            re.split("b([ua]|(s))", "balbulla"))
-        assert ['Hello \udce2\udc9c\udc93', ''] == re.split(r'\r\n|\r|\n',
-                    'Hello \udce2\udc9c\udc93\n')
-
-    def test_weakref(self):
-        import re, _weakref
-        _weakref.ref(re.compile(r""))
-
-    def test_match_compat(self):
-        import re
-        res = re.match(r'(a)|(b)', 'b').start(1)
-        assert res == -1
-
-    def test_pattern_check(self):
-        import _sre
-        raises(TypeError, _sre.compile, {}, 0, [])
-
-    def test_fullmatch(self):
-        import re
-        assert re.compile(r"ab*c").fullmatch("abbcdef") is None
-        assert re.compile(r"ab*c").fullmatch("abbc") is not None
-        assert re.fullmatch(r"ab*c", "abbbcdef") is None
-        assert re.fullmatch(r"ab*c", "abbbc") is not None
-
-    def test_repr(self):
-        import re
-        r = re.compile(r'f(o"\d)', 0)
-        assert repr(r) == (
-            r"""re.compile('f(o"\\d)')""")
-        r = re.compile(r'f(o"\d)', re.IGNORECASE|re.DOTALL|re.VERBOSE)
-        assert repr(r) == (
-            r"""re.compile('f(o"\\d)', re.IGNORECASE|re.DOTALL|re.VERBOSE)""")
-
-    def test_pattern_compare(self):
-        import re
-        pattern1 = re.compile('abc', re.IGNORECASE)
-
-        # equal to itself
-        assert pattern1 == pattern1
-        assert not(pattern1 != pattern1)
-        # equal
-        re.purge()
-        pattern2 = re.compile('abc', re.IGNORECASE)
-        assert hash(pattern2) == hash(pattern1)
-        assert pattern2 == pattern1
-
-        # not equal: different pattern
-        re.purge()
-        pattern3 = re.compile('XYZ', re.IGNORECASE)
-        # warranty that hash values are different
-        assert pattern3 != pattern1
-
-        # not equal: different flag (flags=0)
-        re.purge()
-        pattern4 = re.compile('abc')
-        assert pattern4 != pattern1
-
-        # only == and != comparison operators are supported
-        raises(TypeError, "pattern1 < pattern2")
-
-
 class AppTestSreMatch:
     spaceconfig = dict(usemodules=('array', ))
 
@@ -351,7 +188,7 @@ class AppTestSreMatch:
 
     def test_sub_emptymatch(self):
         import re
-        assert re.sub(r"b*", "*", "abc") == "*a*c*"   # changes in 3.7
+        assert re.sub(r"b*", "*", "abc") == "*a**c*"
 
     def test_sub_shortcut_no_match(self):
         import re
@@ -359,16 +196,12 @@ class AppTestSreMatch:
         assert re.sub(b"a", b"b", s) is s
         s = u"ccccccc"
         assert re.sub(u"a", u"b", s) is s
-        
+
     def test_sub_bytearray(self):
         import re
         assert re.sub(b'a', bytearray(b'A'), b'axa') == b'AxA'
         # this fails on CPython 3.5:
         assert re.sub(b'a', bytearray(b'\\n'), b'axa') == b'\nx\n'
-
-    def test_sub_emptymatch(self):
-        import re
-        assert re.sub(r"b*", "*", "abc") == "*a**c*"
 
     def test_match_array(self):
         import re, array
@@ -443,72 +276,6 @@ class AppTestSreMatch:
         assert matches[1].group() == "b"
 
 
-class AppTestSreScanner:
-
-    spaceconfig = {'usemodules': ['itertools']}
-
-    def test_scanner_attributes(self):
-        import re
-        p = re.compile("bla")
-        s = p.scanner("blablubla")
-        assert p == s.pattern
-
-    def test_scanner_match(self):
-        import re
-        p = re.compile(".").scanner("bla")
-        assert ("b", "l", "a") == (p.match().group(0),
-                                    p.match().group(0), p.match().group(0))
-        assert None == p.match()
-
-    def test_scanner_match_detail(self):
-        import re
-        p = re.compile("a").scanner("aaXaa")
-        assert "a" == p.match().group(0)
-        assert "a" == p.match().group(0)
-        assert None == p.match()
-        # the rest has been changed somewhere between Python 2.6.9
-        # and Python 2.7.18.  PyPy now follows the 2.7.18 behavior
-        assert None == p.match()
-        assert None == p.match()
-
-    def test_scanner_search(self):
-        import re
-        p = re.compile("\d").scanner("bla23c5a")
-        assert ("2", "3", "5") == (p.search().group(0),
-                                    p.search().group(0), p.search().group(0))
-        assert None == p.search()
-
-    def test_scanner_zero_width_match(self):
-        import re, sys
-        if sys.version_info[:2] == (2, 3):
-            skip("2.3 is different here")
-        p = re.compile(".*").scanner("bla")
-        assert ("bla", "") == (p.search().group(0), p.search().group(0))
-        assert None == p.search()
-
-    def test_scanner_empty_match(self):
-        import re, sys
-        p = re.compile("a??").scanner("bac")
-        assert ("", "", "a", "", "") == (
-            p.search().group(0), p.search().group(0), p.search().group(0),
-            p.search().group(0), p.search().group(0))
-        assert None == p.search()
-
-    def test_no_pattern(self):
-        import sre_compile, sre_parse
-        sre_pattern = sre_compile.compile(
-            sre_parse.SubPattern(sre_parse.State()))
-        assert sre_pattern.scanner('s') is not None
-
-    def test_scanner_match_string(self):
-        import re
-        class mystr(type(u"")):
-            pass
-        s = mystr(u"bla")
-        p = re.compile(u".").scanner(s)
-        m = p.match()
-        assert m.string is s
-
 class AppTestGetlower:
     spaceconfig = dict(usemodules=('_locale',))
 
@@ -540,223 +307,6 @@ class AppTestGetlower:
         s.assert_lower_equal([("a", "a"), ("A", "a"), (UPPER_AE, LOWER_AE),
             ("\u00c4", "\u00e4"), (UPPER_PI, LOWER_PI),
             ("\u4444", "\u4444")], sre_constants.SRE_FLAG_UNICODE)
-
-
-class AppTestSimpleSearches:
-
-    spaceconfig = {'usemodules': ('array', 'itertools')}
-
-    def test_search_simple_literal(self):
-        import re
-        assert re.search("bla", "bla")
-        assert re.search("bla", "blab")
-        assert not re.search("bla", "blu")
-
-    def test_search_simple_ats(self):
-        import re
-        assert re.search("^bla", "bla")
-        assert re.search("^bla", "blab")
-        assert not re.search("^bla", "bbla")
-        assert re.search("bla$", "abla")
-        assert re.search("bla$", "bla\n")
-        assert not re.search("bla$", "blaa")
-
-    def test_search_simple_boundaries(self):
-        import re
-        UPPER_PI = "\u03a0"
-        assert re.search(r"bla\b", "bla")
-        assert re.search(r"bla\b", "bla ja")
-        assert re.search(r"bla\b", "bla%s" % UPPER_PI, re.ASCII)
-        assert not re.search(r"bla\b", "blano")
-        assert not re.search(r"bla\b", "bla%s" % UPPER_PI, re.UNICODE)
-
-    def test_search_simple_categories(self):
-        import re
-        LOWER_PI = "\u03c0"
-        INDIAN_DIGIT = "\u0966"
-        EM_SPACE = "\u2001"
-        LOWER_AE = "\xe4"
-        assert re.search(r"bla\d\s\w", "bla3 b")
-        assert re.search(r"b\d", "b%s" % INDIAN_DIGIT, re.UNICODE)
-        assert not re.search(r"b\D", "b%s" % INDIAN_DIGIT, re.UNICODE)
-        assert re.search(r"b\s", "b%s" % EM_SPACE, re.UNICODE)
-        assert not re.search(r"b\S", "b%s" % EM_SPACE, re.UNICODE)
-        assert re.search(r"b\w", "b%s" % LOWER_PI, re.UNICODE)
-        assert not re.search(r"b\W", "b%s" % LOWER_PI, re.UNICODE)
-        assert re.search(r"b\w", "b%s" % LOWER_AE, re.UNICODE)
-
-    def test_search_simple_any(self):
-        import re
-        assert re.search(r"b..a", "jboaas")
-        assert not re.search(r"b..a", "jbo\nas")
-        assert re.search(r"b..a", "jbo\nas", re.DOTALL)
-
-    def test_search_simple_in(self):
-        import re
-        UPPER_PI = "\u03a0"
-        LOWER_PI = "\u03c0"
-        EM_SPACE = "\u2001"
-        LINE_SEP = "\u2028"
-        assert re.search(r"b[\da-z]a", "bb1a")
-        assert re.search(r"b[\da-z]a", "bbsa")
-        assert not re.search(r"b[\da-z]a", "bbSa")
-        assert re.search(r"b[^okd]a", "bsa")
-        assert not re.search(r"b[^okd]a", "bda")
-        assert re.search("b[%s%s%s]a" % (LOWER_PI, UPPER_PI, EM_SPACE),
-            "b%sa" % UPPER_PI) # bigcharset
-        assert re.search("b[%s%s%s]a" % (LOWER_PI, UPPER_PI, EM_SPACE),
-            "b%sa" % EM_SPACE)
-        assert not re.search("b[%s%s%s]a" % (LOWER_PI, UPPER_PI, EM_SPACE),
-            "b%sa" % LINE_SEP)
-
-    def test_search_simple_literal_ignore(self):
-        import re
-        UPPER_PI = "\u03a0"
-        LOWER_PI = "\u03c0"
-        assert re.search(r"ba", "ba", re.IGNORECASE)
-        assert re.search(r"ba", "BA", re.IGNORECASE)
-        assert re.search("b%s" % UPPER_PI, "B%s" % LOWER_PI,
-            re.IGNORECASE | re.UNICODE)
-
-    def test_search_simple_in_ignore(self):
-        import re
-        UPPER_PI = "\u03a0"
-        LOWER_PI = "\u03c0"
-        assert re.search(r"ba[A-C]", "bac", re.IGNORECASE)
-        assert re.search(r"ba[a-c]", "baB", re.IGNORECASE)
-        assert re.search("ba[%s]" % UPPER_PI, "ba%s" % LOWER_PI,
-            re.IGNORECASE | re.UNICODE)
-        assert re.search(r"ba[^A-C]", "bar", re.IGNORECASE)
-        assert not re.search(r"ba[^A-C]", "baA", re.IGNORECASE)
-        assert not re.search(r"ba[^A-C]", "baa", re.IGNORECASE)
-
-    def test_search_simple_branch(self):
-        import re
-        assert re.search(r"a(bb|d[ef])b", "adeb")
-        assert re.search(r"a(bb|d[ef])b", "abbb")
-
-    def test_search_simple_repeat_one(self):
-        import re
-        assert re.search(r"aa+", "aa") # empty tail
-        assert re.search(r"aa+ab", "aaaab") # backtracking
-        assert re.search(r"aa*ab", "aab") # empty match
-        assert re.search(r"a[bc]+", "abbccb")
-        assert "abbcb" == re.search(r"a.+b", "abbcb\nb").group()
-        assert "abbcb\nb" == re.search(r"a.+b", "abbcb\nb", re.DOTALL).group()
-        assert re.search(r"ab+c", "aBbBbBc", re.IGNORECASE)
-        assert not re.search(r"aa{2,3}", "aa") # string too short
-        assert not re.search(r"aa{2,3}b", "aab") # too few repetitions
-        assert not re.search(r"aa+b", "aaaac") # tail doesn't match
-
-    def test_search_simple_min_repeat_one(self):
-        import re
-        assert re.search(r"aa+?", "aa") # empty tail
-        assert re.search(r"aa+?ab", "aaaab") # forward tracking
-        assert re.search(r"a[bc]+?", "abbccb")
-        assert "abb" == re.search(r"a.+?b", "abbcb\nb").group()
-        assert "a\nbb" == re.search(r"a.+b", "a\nbbc", re.DOTALL).group()
-        assert re.search(r"ab+?c", "aBbBbBc", re.IGNORECASE)
-        assert not re.search(r"aa+?", "a") # string too short
-        assert not re.search(r"aa{2,3}?b", "aab") # too few repetitions
-        assert not re.search(r"aa+?b", "aaaac") # tail doesn't match
-        assert re.match(".*?cd", "abcabcde").end(0) == 7
-
-    def test_search_simple_repeat_maximizing(self):
-        import re
-        assert not re.search(r"(ab){3,5}", "abab")
-        assert not re.search(r"(ab){3,5}", "ababa")
-        assert re.search(r"(ab){3,5}", "ababab")
-        assert re.search(r"(ab){3,5}", "abababababab").end(0) == 10
-        assert "ad" == re.search(r"(a.)*", "abacad").group(1)
-        assert ("abcg", "cg") == (
-            re.search(r"(ab(c.)*)+", "ababcecfabcg").groups())
-        assert ("cg", "cg") == (
-            re.search(r"(ab|(c.))+", "abcg").groups())
-        assert ("ab", "cf") == (
-            re.search(r"((c.)|ab)+", "cfab").groups())
-        assert re.search(r".*", "")
-
-    def test_search_simple_repeat_minimizing(self):
-        import re
-        assert not re.search(r"(ab){3,5}?", "abab")
-        assert re.search(r"(ab){3,5}?", "ababab")
-        assert re.search(r"b(a){3,5}?b", "baaaaab")
-        assert not re.search(r"b(a){3,5}?b", "baaaaaab")
-        assert re.search(r"a(b(.)+?)*", "abdbebb")
-
-    def test_search_simple_groupref(self):
-        import re
-        UPPER_PI = "\u03a0"
-        LOWER_PI = "\u03c0"
-        assert re.match(r"((ab)+)c\1", "ababcabab")
-        assert not re.match(r"((ab)+)c\1", "ababcab")
-        assert not re.search(r"(a|(b))\2", "aa")
-        assert re.match(r"((ab)+)c\1", "aBAbcAbaB", re.IGNORECASE)
-        assert re.match(r"((a.)+)c\1", "a%sca%s" % (UPPER_PI, LOWER_PI),
-            re.IGNORECASE | re.UNICODE)
-
-    def test_search_simple_groupref_exists(self):
-        import re, sys
-        if not sys.version_info[:2] == (2, 3):
-            assert re.search(r"(<)?bla(?(1)>)", "<bla>")
-            assert re.search(r"(<)?bla(?(1)>)", "bla")
-            assert not re.match(r"(<)?bla(?(1)>)", "<bla")
-            assert re.search(r"(<)?bla(?(1)>|u)", "blau")
-
-    def test_search_simple_assert(self):
-        import re
-        assert re.search(r"b(?=\d\d).{3,}", "b23a")
-        assert not re.search(r"b(?=\d\d).{3,}", "b2aa")
-        assert re.search(r"b(?<=\d.)a", "2ba")
-        assert not re.search(r"b(?<=\d.)a", "ba")
-
-    def test_search_simple_assert_not(self):
-        import re
-        assert re.search(r"b(?<!\d.)a", "aba")
-        assert re.search(r"b(?<!\d.)a", "ba")
-        assert not re.search(r"b(?<!\d.)a", "11ba")
-
-
-class AppTestMarksStack:
-
-    spaceconfig = {'usemodules': ['itertools']}
-
-    def test_mark_stack_branch(self):
-        import re
-        m = re.match("b(.)a|b.b", "bob")
-        assert None == m.group(1)
-        assert None == m.lastindex
-
-    def test_mark_stack_repeat_one(self):
-        import re
-        m = re.match("\d+1((2)|(3))4", "2212413")
-        assert ("2", "2", None) == m.group(1, 2, 3)
-        assert 1 == m.lastindex
-
-    def test_mark_stack_min_repeat_one(self):
-        import re
-        m = re.match("\d+?1((2)|(3))44", "221341244")
-        assert ("2", "2", None) == m.group(1, 2, 3)
-        assert 1 == m.lastindex
-
-    def test_mark_stack_max_until(self):
-        import re
-        m = re.match("(\d)+1((2)|(3))4", "2212413")
-        assert ("2", "2", None) == m.group(2, 3, 4)
-        assert 2 == m.lastindex
-
-    def test_mark_stack_min_until(self):
-        import re
-        m = re.match("(\d)+?1((2)|(3))44", "221341244")
-        assert ("2", "2", None) == m.group(2, 3, 4)
-        assert 2 == m.lastindex
-
-    def test_bug_725149(self):
-        # mark_stack_base restoring before restoring marks
-        # test copied from CPython test
-        import re
-        assert re.match('(a)(?:(?=(b)*)c)*', 'abb').groups() == ('a', None)
-        assert re.match('(a)((?!(b)*))*', 'abb').groups() == ('a', None, None)
 
 
 class AppTestOpcodes:
@@ -1103,125 +653,3 @@ class AppTestOpcodes:
         assert "a" == s.search(opcodes, "ac").group(0)
         s.assert_match(opcodes, ["a"])
         s.assert_no_match(opcodes, ["ab"])
-
-
-class AppTestOptimizations:
-    """These tests try to trigger optmized edge cases."""
-
-    spaceconfig = {'usemodules': ['itertools']}
-
-    def test_match_length_optimization(self):
-        import re
-        assert None == re.match("bla", "blub")
-
-    def test_fast_search(self):
-        import re
-        assert None == re.search("bl", "abaub")
-        assert None == re.search("bl", "b")
-        assert ["bl", "bl"] == re.findall("bl", "blbl")
-        assert ["a", "u"] == re.findall("bl(.)", "blablu")
-
-    def test_branch_literal_shortcut(self):
-        import re
-        assert None == re.search("bl|a|c", "hello")
-
-    def test_literal_search(self):
-        import re
-        assert re.search("b(\d)", "ababbbab1")
-        assert None == re.search("b(\d)", "ababbbab")
-
-    def test_repeat_one_literal_tail(self):
-        import re
-        assert re.search(".+ab", "wowowowawoabwowo")
-        assert None == re.search(".+ab", "wowowaowowo")
-
-    def test_split_empty(self):
-        import re
-        assert re.split('', '') == ['', '']
-        assert re.split('', 'ab') == ['', 'a', 'b', '']
-        assert re.split('a*', '') == ['', '']
-        assert re.split('a*', 'a') == ['', '', '']
-        assert re.split('a*', 'aa') == ['', '', '']
-        assert re.split('a*', 'baaac') == ['', 'b', '', 'c', '']
-
-    def test_type_names(self):
-        import re
-        assert repr(re.Pattern) == "<class 're.Pattern'>"
-        assert repr(re.Match) == "<class 're.Match'>"
-
-class AppTestUnicodeExtra:
-    def test_string_attribute(self):
-        import re
-        match = re.search(u"\u1234", u"\u1233\u1234\u1235")
-        assert match.string == u"\u1233\u1234\u1235"
-        # check ascii version too
-        match = re.search(u"a", u"bac")
-        assert match.string == u"bac"
-
-    def test_string_attribute_is_original_string(self):
-        import re
-        class mystr(type(u"")):
-            pass
-        s = mystr(u"\u1233\u1234\u1235")
-        match = re.search(u"\u1234", s)
-        assert match.string is s
-
-    def test_match_start(self):
-        import re
-        match = re.search(u"\u1234", u"\u1233\u1234\u1235")
-        assert match.start() == 1
-
-    def test_match_repr_span(self):
-        import re
-        match = re.match(u"\u1234", u"\u1234")
-        assert match.span() == (0, 1)
-        assert "span=(0, 1), match='\u1234'" in repr(match)
-
-    def test_match_repr_truncation(self):
-        import re
-        s = "xy" + u"\u1234" * 50
-        match = re.match(s, s)
-        # this used to produce invalid utf-8 by truncating repr(s)
-        # after 50 bytes
-        assert "span=(0, 52), match=" + repr(s)[:50] + ">" in repr(match)
-
-    def test_pattern_repr_truncation(self):
-        import re
-        s = "xy" + u"\u1234" * 200
-        pattern = re.compile(s)
-        # this used to produce invalid utf-8 by truncating repr(s)
-        # after 200 bytes
-        assert repr(pattern) == "re.compile(%s)" % (repr(s)[:200],)
-
-    def test_subx_unusual_types_no_match(self):
-        # issue #3515
-        import re
-        result = re.sub(b"x", lambda x: 1/0, memoryview(b"yz"))
-        assert type(result) is bytes
-        assert result == b"yz"
-        class U(str): pass
-        result = re.sub(u"x", lambda x: 1/0, U(u"yz"))
-        assert type(result) is str
-        assert result == u"yz"
-        class B(bytes): pass
-        result = re.sub(b"x", lambda x: 1/0, B(b"yz"))
-        assert type(result) is bytes
-        assert result == b"yz"
-
-    def test_bug_40736(self):
-        # Passing a non-string/bytes-like object to re.search() must raise
-        # TypeError with a message mentioning the actual type, not a misleading
-        # "can't use a string pattern on a bytes-like object" message.
-        import re
-        try:
-            re.search("x*", 5)
-        except TypeError as e:
-            assert "got 'int'" in str(e), str(e)
-        else:
-            raise AssertionError("expected TypeError")
-        try:
-            re.search("x*", None)
-        except TypeError as e:
-            assert "got 'NoneType'" in str(e), str(e)
-        else:
-            raise AssertionError("expected TypeError")
